tmpl: fix rendering issue
When rendering using raw `hugo`, the function `page`, expected to return
the current `Page` object returns `nil` in partials called in
shortcodes, and page.RenderString is needed for rendering callout
labels. So we use the function `site` to get the `Site` object and find
whichever a `Page` and use its RenderString function.

fixes: 957c5e0