docs: sync libLoL design doc with loongson-community/areweloongyet@0db1dda844bf
Previously libLoL has no project page (this repo), so its documentation
was originally developed over at AWLY. Now it's probably best for
project-specific details and docs to live in the project's own website.

The previous design principle doc was an early revision, so first bring
it up-to-date with the AWLY version. Editorial changes will come later.